chore: move everything about actual preview to RemoveMeshPreviewController
